requirements.txt
pandas
numpy
scikit-learn
joblib
python-telegram-bot==20.3